+++ src/lib/sgzlib.h
@@ -137,15 +137,16 @@
 #include <sys/stat.h>
 #include <unistd.h>
 #include <stdarg.h>
+#include <stdint.h>
 
 /* Header to be written at the start of the file */
 struct sgzip_header {
   char magic[3];
-  unsigned int blocksize;
+  uint32_t blocksize;
   union {
     char compression[4];
     //After being loaded from the file, this will be where we store the 
-    long int max_chunks;
+    uint32_t max_chunks;
   } x;
 }  __attribute__((packed));
 
@@ -157,7 +158,7 @@ struct sgzip_obj {
 
 /* This linked list stores the index as we are building the file */
 struct sgzip_index_list {
-  unsigned long long int offset;
+  uint64_t offset;
   struct sgzip_index_list *next;
 }__attribute__((packed));
 
@@ -178,8 +179,8 @@ struct sgzip_header *sgzip_write_header(int fd,struct sgzip_header *header);
 void sgzip_compress_fds(int infd,int outfd,const struct sgzip_obj *obj);
 
 /* read a random buffer from the sgziped file */
-int sgzip_read_random(char *buf, int len, unsigned long long int offs,
-		int fd, unsigned long long int *index,const struct sgzip_obj *obj);
+int sgzip_read_random(char *buf, int len, uint64_t offs,
+		int fd, uint64_t *index,const struct sgzip_obj *obj);
 
 /* 
    Reads the index from the file and returns an array of long ints
@@ -191,7 +192,7 @@ int sgzip_read_random(char *buf, int len, unsigned long long int offs,
 
    If the index is not there we flag an error by returning null.
 */
-unsigned long long int *sgzip_read_index(int fd,struct sgzip_obj *sgzip) ;
+uint64_t *sgzip_read_index(int fd,struct sgzip_obj *sgzip) ;
 
 /* 
    reads the stream and calculates the index map.
@@ -200,11 +201,11 @@ unsigned long long int *sgzip_read_index(int fd,struct sgzip_obj *sgzip) ;
    rebuilding the index. None of the blocks are decoded so this should be quick.
    We return an index_list list.
  */
-unsigned long long int *sgzip_calculate_index_from_stream(int fd,const struct sgzip_obj *obj);
+uint64_t *sgzip_calculate_index_from_stream(int fd,const struct sgzip_obj *obj);
 
 /* Write the index passed in at the current positon in the file. Note
    you would probably only want to do this after
    sgzip_calculate_index_from_stream so that the correct position is
    in the file */
-void sgzip_write_index(int outfd,unsigned long long int *index) ;
+void sgzip_write_index(int outfd,uint64_t *index) ;
 void sgzip_decompress_fds(int fd,int outfd,struct sgzip_obj *sgzip);
